+++ NEWS.md
@@ -16,4 +16,4 @@
 
 # epmfd 0.2.0
 
-* Created the package "epmfd" version 0.2.0
+-   Created the package "epmfd" version 0.2.0